Updated the changelog with all of the changes made so far.
In order to prevent myself from scanning a ton of commits before each
update (of which v0.2.0 will be huge), I'm going to try and make a habit
of updating the changelog as I am developing the game.

And before you ask, no this is not me trying to put off writing unit
tests... no, totally not. Not at all! I'm not scared, you are.