Added functions to set up scene entries to AssetPackTypeCatalog.
This involved the creation of the GeoData class, which holds metadata
about the geometry of custom scenes (average vertex, bounding box).
This is pretty much the same as in v0.1.x, the only difference being
the metadata is now stored as a `.geo` file rather than a `.box` file
since the file now stores more data that just the bounding box.