Finished the AssetPackCatalog class.
- Added the AssetEntrySave class, which is used to represent pre-made
save files that can be used to load games.

- Added the CustomValue.get_value_variant() function.

- Added the AssetPackTypeCatalog.write_entry_to_config() function,
which is essentially the reverse of apply_config_to_entry, in that
it takes an already existing entry and writes the config section
required to obtain that entry's properties. This is used to load a
parent entry's properties so that they can then be overwritten and
applied to a new entry, but this function will also come in handy
in the future once we start creating asset packs from within the
game.