.md5 files are now removed in remove_untagged(), not untag().
Previously, if one wanted to remove the .md5 file associated with a
rogue file, then they would have to use untag() on it. However, that
required that the file was already tagged, which is not guaranteed.
Tagging, then untagging, a file to remove its metadata is not a great
solution, so now all .md5 files associated with rogue files are now
removed at once in remove_untagged().

This now means that in AssetPackCatalog, the cleaning function will
now remove the stored metadata for files that were previously imported,
but no longer appear in the type directory. Then, if the files were to
appear again, the directory would tag them as new instead of being
unchanged, which is the intended functionality.