Added the ImportProgressPanel scene.
Finally... after all of this work re-writing the import system...
I can actually SEE the import code in action (and not in a test)!

\o/

This works by emitting signals from each of the catalog classes up
the chain, until they get to AssetCatalogInteractive, where the
current progress is stored in a thread-safe manner. At any time, the
main thread can ask for a copy of the progress report, and display
it in the UI.

This now means that the player does not have to wait for all of the
custom assets to be imported before the game launches, instead, the
game will go straight to loading the main menu, and once it is
running, the importing process will start independently of the rest
of the game. Only when the process is complete is the AssetDB
modified, and the rest of the game notified of the new assets.

The UI showing the import progress has also been improved. It is now
a panel, with a loading circle and two progress bars: one for the
packs, and one for the files in each of those packs. Once the
process is complete, it automatically hides itself.