Moved boards into the Godot project.
There were two things I noticed when the engine imported the files:

- The engine automatically set the import variables for textures
that were a part of meshes to a 3D-specific preset. Because of this,
I am considering going through every texture in the project and
setting the import variables appropiately for each one in order to
optimise the game's video memory usage.

- The engine would not assign the textures to the mesh's materials
if the file name came after the mesh's. I have added a workaround by
prepending a '00_' to the texture filenames.