Transferred containers into the Godot project from the asset pack.
This was a little tricker since two of the objects were glTF files,
one of which transformed the object several times over a number of
nodes. This was solved by modifying the files to only contain one
node, and for the imported object to be transformed via its parent
collision shape in the final scene.