Fixed text templates not being loaded correctly within the project.
This was because text files aren't actually imported as resources
in vanilla Godot, so if an exported variable is looking for a path
to a resource, and it points to a text file, it won't actually set
the variable when the resource is loaded.

This has been fixed by creating a small editor plugin whose only
job is to import text files like regular resources.

Resources that used the ActualTextFile class have been converted to
regular text files, and other text files that are in the project
directory structure, but aren't actually used by the project itself,
are now ignored with '.gdignore' files.