Simplified scene switching behaviour, moved loading scene.
The plan is to have the main menu and the game room in the same
scene as to avoid loading screens when entering a game, so it didn't
make sense to have the extra functionality previously in the Global
class.

Now, SceneTree.change_scene_to() is used in all instances, from
SplashScreen->LoadingScreen, and from LoadingScreen->Game (currently
ImportProgressPanel, as I haven't made the new Game scene yet).

- Increased the font size of labels in the loading scene.