Added the IndexManager node.
This node was added as an optimisation for nodes whose children are
named as integers, e.g. the PieceManager. Instead of the engine
checking one-by-one each child for the one with a given name,
each child is added into an array such that its index in the array
matches its given index/name.

Future nodes such as the PieceManager and HiddenAreaManager will be
derived from this new class.