Started work on the new main menu screen.
- Icons have been added to each of the buttons.
- The Singleplayer and Multiplayer buttons are the biggest, followed
by Collection, Options, and Credits, and then the Links, Info, and
Quit buttons.
- The main menu gradually fades in when it is first shown.

Since in 4.x the Popup controls are actually in separate windows
to the main one, I'm planning on overhauling how popups are displayed
in the main menu, most likely with a PanelPopup, as well as hiding the
main controls when they appear.