Re-implemented the third person camera.
The code that handles player input and moves the camera has not
changed much by design, so that the controls feel consistent between
versions. This commit is mostly re-factoring.

However, controller input is now registered and used by the game to
move the camera, as to fulfill #100.

The new ThirdPersonCamera scene is now a subclass of CameraController,
since I also plan to add a first person camera in the future.
Unlike v0.1.x, the CameraController scripts will only serve to move
the camera. The PlayerController will be the main player input script,
which will handle the manipulation of pieces, etc.