Added test for loading v0.1.x state files.
A couple of things have come up from this process that will need to
be solved in future commits:

- Tables need to adjust their centre-of-mass, and thus their position
in the scene.

- There is currently no way to distinguish if a custom colour needs
to be applied to a piece or not.

- There might be some discrepencies between the scale of objects in
v0.1.x and v0.2.0, depending on how the piece scene is made in the
new version. Custom fixes may need to be applied to these pieces.