Started on the new PaintPlane and PaintViewport.
The PaintViewport is what holds the paint image and modifies it.
Instead of using a shader to draw over the image and then manually
saving the viewport texture each frame, the clear mode is changed
depending on what operation is happening. For example, if the image
is being set from loading a save file, the viewport is cleared for
that one frame, then a sprite is used to display the image within
the viewport. Subsequent paint commands would not clear the viewport,
so that they paint on top of what is already there.

The PaintPlane is what displays the PaintViewport in 3D space.
It is attatched to the table so when the table is flipped, the plane
will follow. The reason it is a class now is to avoid errors being
thrown during the tree setup when it was a scene.