Converted default pack save games from *.tc files to *.tres
Since the save files are now in text format rather than binary, it
will make them much easier to version control. They also no longer
hold any version information, so they do not need to be updated
before every release.

However, the Go save file is much larger than it needs to be, since
there are 360 stones in their containers, each of which need their
own unique index. If this requirement goes away in the future, or
if there is a way to make it so containers have an infinite amount
of a certain object, then we can optimise the size of this file.