Fixed margins resetting in the AttentionPanel scenes.
This was due to the panel itself setting the margins of it child
controls to (4, 4, -4, -4). For scenes where this was desired,
the MarginContainer has been removed as it is no longer needed.

For other scenes, the MarginContainer has been kept, but the top
and bottom margins are no longer modified, and the left and right
margins have been adjusted to account for it's own margins being
set by the parent panel.