Started work on the GameConfig class.
This will effectively be the back-end for the game's settings, and
it will be what the options menu interfaces with.

This allows for most of the game's settings to be loaded and applied
during the splash screen without having to load the entire options
menu scene.

Also, Happy New Year! \o/