Moved the section buttons to the top of the options menu.
This is because when I was navigating the menu with the section
buttons on the left, using just keyboard controls, when the focus
was on a slider (which will be quite common) pressing the left key
moved the slider as intended, but it would never move focus to the
section button.

So either the player would have had to keep pressing down or up to
move away from that section and shift the focus, which doesn't make
sense in terms of the direction the focus goes, or have the "ui_back"
input go back to the section buttons. I opted to go back to a similar
layout to v0.1.x, which makes the focus directions more intuitive.

- In order for the text to fit inside the button, and to make the
section more generic, the "Multiplayer" section has been renamed
to just "Player".