Added the remaining general_* properties to the options menu.
They all work well on keyboard and mouse, but the spin box for the
autosave file count doesn't work well with controller inputs.
In the future I'm going to make a custom class that acts like a spin
box, but has a - and + button on either side. This can then be used
elsewhere in the UI once it is complete.