Created the new IntegerSpinBox control to replace SpinBox.
The issue with the vanilla spin box is that you cannot change its
value with just the controller - you need keyboard inputs to change
the text in the internal LineEdit.

This new version has an increment and decrement button on either side
that is focusable with the controller, and a label in the middle
showing the current value. If the label is clicked, then it is
replaced with a LineEdit that is editable with the keyboard.