Added the multiplayer_* properties to the new options menu.
However, the multiplayer_color property has no control yet. This is
because I'm planning to make my own control (again) that is controller
friendly.

- AttentionPanels now close when the "ui_cancel" input is pressed,
not released. This is so pressing "ui_cancel" while in a control's
pop-up menu will just close the pop-up menu, not the entire panel.

- The multiplayer_chat_font_size options has been moved to the
"General" tab, since I think it makes more sense for it to be there
alongside the "Show System Warnings" and "Show System Errors" options.

- Players will now have their own styled menu buttons in the player
list, and a preview has been added to the "Player" tab.