Added a new tool to replace ColorButton: ColorSlider!
It has two modes: HSV and RGB, and each mode has a set of three
sliders that can be used to change the colour, which is shown at the
bottom of the control along with some information about the colour.

Best of all, it can be used with a controller! \o/