Fixed the main menu being hidden while a popup was in view.
This was caused by the fact that on controller, there are now two
different buttons for closing menus depending on the menu, B/O for
closing popups, and the Start button for closing the main menu.
So the player could open the options menu for example, and then
press the Start button to close the main menu, but the options menu
would remain visible.

This has been fixed by having the main script check if the main menu
is currently showing a popup to the player before hiding it.