The "ui_cancel" input now also closes the main menu when in-game.
Previously on controller, when navigating from the options menu back
to the game, one would have to press B/O, then the Start button to
close the main menu, as B/O did not work.

Now, like with the ESC key, the player can just press B/O repeatedly
to get back to the game.