The D-Pad on controllers can now emit echo input events.
This makes using certain controls like sliders much nicer on controller,
since the directional button can just be held down to make the grabber
move faster, just like on keyboard.

However, this doesn't work when the scene tree has been paused, which
is what happens when the main menu becomes visible. So we'll need to
find another way to disable input to the player camera while the menu
is open.