Added the GameConfig.video_ui_scale property.
This property allows the UI to be scaled between 50% and 125% of it's
original size. 50% was the point at which text started to become
illegible on my monitor, and past 125% the main menu started to go off
the bottom of the screen. Some tweaking will be necessary in order to
allow for > 125% scaling in the future.

This commit solves #290, but the slider will be added to the options
menu at a later date.