Fixed some bindings being removed after restarting the game.
This was due to the way the GameConfig read the config file: If the
first binding of an action was the default, it is not kept in the file,
so it was not added to an array. But if the second binding was changed,
then that binding, and only that binding, was added to the array.

This was fixed by adding the default bindings to the array if the
section-key did not exist in the config file.