Added Message, MessageHub, ChatTextLabel classes.
I've started working on the new chat window. It will stay mostly the
same as before, with two major differences:

- Instead of generating BBCode text and appending it each time a
message is received, the RichTextLabel's push and pop functions are
used instead.

- Instead of the chat window being the central location where RPCs
are exchanged when players send each other messages, the chat window
will just show messages that are sent to the MessageHub. The RPCs will
be moved to the MessageHub to decouple it from the UI.