Organised how the UI focus interacts with the chat window.
- Added two new bindings: 'game_chat', and 'game_command', for
starting a message and a command respectively.

- If the Game node detects an unhandled mouse click, any controls
that have focus release that focus. This is mostly for releasing the
focus from the chat window when the player wants to either drag an
object, or rotate the camera.

- Organised the chat window's focus neighbours such that certain
directions will get the focus to leave the chat window, and either
lose it indefinitely (while in-game), or to give it back to another
control (while the main menu is visible).

This now means that even though the chat window and the main menu
have no common parent node that is a Control (the only common node
is the Game node, which is just a regular node), there is now a way
for the focus to be transferred between them.

From the main menu (or in-game) to the chat window, either Enter or
'/' can be pressed to give the focus to the ChatLineEdit, from which
the focus can be moved to the other elements of the ChatWindow.

NOTE: The Enter key cannot be used while in the main menu, as this key
is also used for 'ui_accept', and would clash otherwise.

From the ChatWindow, the focus can be transferred back to the main
menu by navigating the focus away from the window.