Started work on the new MasterServer class.
The plan is to mostly keep the MasterServer class the same, just with
some tidying up here and there. However, the code responsible for
establishing connections with other peers will be moved from the
main game script, to a new class called NetworkManager.

I'm making a commit now since I realised that this branch does not
have the new SSL certificate for the master server yet... oops.