Client now checks if connection to the master server has been lost.
Using a patch to the engine that provides a new signal, that being
WebSocketClient.ping_received, the MasterServer class now uses a
timer which is reset every time a ping is received from the server,
in order to make sure that the connection is still alive.

If a ping has not been received in a given amount of time (by default
30 seconds), then the connection is assumed to be lost, and the
'connection_lost' signal is fired.

NOTE: You will need to download and compile the latest commit of the
tabletop-3.5.3 branch of the official fork for this to work properly.