Created the NetworkManager class.
This class is used to create and maintain the network peers, and to
provide an easier-to-understand interface within the networking
stack, of which there will be a Lobby class using this soon!

Still got some things left to figure out, e.g. sealed rooms, and
handling master server disconnects, but it's very close! I would
work on it more, but it's just gone 6am, and I'm very much in
need of sleep ><

- Clients no longer attempt to establish connections with other
clients, as they do not send packets to each other.

- Updated the WebRTC library from v0.5 to v1.0.5!

This commit should hopefully deal with #282, but I will need help
with testing later on once the UI is done.