Started work on MultiplayerMainPanel, MultiplayerSetupPanel.
A new panel has been added for setting up the network. The controls
that are shown change depending on if the player wants to host, or
to join a game, and whether they want to use a room code or an
IP address.

This commit starts work on #225 - the new "Direct Connection" mode
will use the ENet protocol instead of WebRTC. Implementation of this
new mode will be added to the NetworkManager in a future commit.