Added close code 4018: JOIN_QUEUE_FULL.
This close code won't be given by the current master server, but the
re-written one that I'm working on in Rust that I plan to use in the
future.

Instead of having a maximum number of players across all rooms, instead
the new master server will have a maximum number of players in each
room, e.g. 10, and on top of that there is a maximum number of players
that have connected, but have yet to create or join a room. This is
called the "join queue".