Started work on the DataReceiver node.
This node will allow clients to send byte arrays of any size to the
server, across multiple packets.

This will be how clients send new states to the server - the server
will then use the future DataBroadcaster class to send this data to
all of the other clients, with syncing capabilities.