Added the StateFreeze system.
This will be used to freeze the current state of the game on all
clients before transferring states between them - by freezing, this
means that clients will not be able to move the camera or any objects,
and they cannot manipulate the state in any way.

Once all of the clients have acknowledged the signal from the server,
a secondary signal is fired on the server which can then be used in
the future to initiate the transfer.